# docs/api-specification.yaml
...
openapi: 3.0.3
info:
  title: RocketVault API
  description: |
    A self-hosted, open-source alternative to Azure Key Vault: secrets
    management, cryptographic key operations, X.509 certificate lifecycle
    management, and multi-vault RBAC.
  version: 4.0.0
  contact:
    name: RocketVault
    url: https://github.com/Snehal1112/rocketvault
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.rocketvault.local
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/vaults/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: Name of the vault
        schema:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$"
          example: "production"
    get:
      summary: Get a vault
      description: |
        Returns a vault by name. Authorized against the named vault itself, not
        the default vault.
      operationId: getVault
      tags:
        - Vaults
      responses:
        "200":
          description: Vault retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Update a vault
      description: |
        Updates a vault's enabled flag, purge protection, retention period, or
        tags. Nil fields are left unchanged; an empty tag map clears all tags.
      operationId: updateVault
      tags:
        - Vaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultRequest"
      responses:
        "200":
          description: Vault updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Soft-delete a vault
      description: |
        Soft-deletes a vault and cascades to its secrets, keys, and
        certificates. The default vault cannot be deleted.
      operationId: deleteVault
      tags:
        - Vaults
      responses:
        "204":
          description: Vault soft-deleted successfully
        "400":
          description: The default vault is protected from deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/vaults/{name}/webhook:
    parameters:
      - name: name
        in: path
        required: true
        description: Name of the vault
        schema:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$"
          example: "production"
    put:
      summary: Create or update a vault's webhook config
      description: |
        Creates or updates the vault's webhook notification target.
        Authorized against the named vault itself (`vaults/manage`), the same
        as the other `/vaults/{name}` management routes.

        The server always mints the signing secret; there is no
        client-supplied secret field. `rotate_secret: true` replaces an
        existing secret. `enabled` is omittable to leave the current value
        (or default to `true` on create) rather than silently disabling the
        webhook by omission.

        The response carries `signing_secret` in plaintext only on the call
        that mints one — the first create, or any update with
        `rotate_secret: true`. It is shown exactly once: the secret is stored
        encrypted, and no endpoint (including the GET below) ever returns it
        again after this response.
      operationId: upsertVaultWebhook
      tags:
        - Vaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVaultWebhookRequest"
      responses:
        "200":
          description: |
            Webhook config saved. The `signing_secret` field is present only
            when this call minted a secret (create, or update with
            `rotate_secret: true`); an update that did not rotate returns the
            config without it.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/VaultWebhookConfigCreatedResponse"
                  - $ref: "#/components/schemas/VaultWebhookConfigResponse"
              examples:
                created_or_rotated:
                  summary: Create or rotate — secret shown once
                  value:
                    url: "https://hooks.example.com/rocketvault"
                    enabled: true
                    created_at: "2026-08-19T12:00:00Z"
                    updated_at: "2026-08-19T12:00:00Z"
                    signing_secret: "b64-encoded-secret-shown-only-now"
                updated_without_rotate:
                  summary: Update without rotating — no secret field
                  value:
                    url: "https://hooks.example.com/rocketvault"
                    enabled: true
                    created_at: "2026-08-19T12:00:00Z"
                    updated_at: "2026-08-19T12:05:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: Get a vault's webhook config
      description: |
        Returns the vault's webhook config. The response never includes the
        signing secret in any form, encrypted or plain — the secret is shown
        exactly once, from the PUT call that minted it.
      operationId: getVaultWebhook
      tags:
        - Vaults
      responses:
        "200":
          description: Webhook config retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultWebhookConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The vault does not exist, or has no webhook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a vault's webhook config
      description: Removes the vault's webhook config. Deleting when none exists succeeds.
      operationId: deleteVaultWebhook
      tags:
        - Vaults
      responses:
        "204":
          description: Webhook config deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/v1/users/login`,
        `POST /api/v1/users/refresh`, the OIDC callback, or
        `POST /api/v1/oauth2/token`. Tokens are signed asymmetrically
        (RS256/ES256) and always carry a `kid` header.
...
  responses:
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or the session is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The caller is not authorized for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - id
        - message
        - status_code
      properties:
        id:
          type: string
          description: Error identifier
          example: "api.secret.create"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request body"
        detailed_error:
          type: string
          description: Detailed error information
          example: "json: invalid character '}' looking for beginning of value"
        status_code:
          type: integer
          description: HTTP status code
          example: 400
...
    UpdateVaultRequest:
      type: object
      description: Omitted fields are left unchanged.
      properties:
        enabled:
          type: boolean
        purge_protection:
          type: boolean
        retention_days:
          type: integer
        tags:
          type: object
          description: Omitted leaves tags unchanged, `{}` clears them, a value replaces them
          additionalProperties:
            type: string
    VaultResponse:
      type: object
      required:
        - id
        - name
        - enabled
        - purge_protection
        - retention_days
        - created_by
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        purge_protection:
          type: boolean
        retention_days:
          type: integer
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
        scheduled_purge_at:
          type: string
          format: date-time
        tags:
          type: object
          additionalProperties:
            type: string
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
          format: uuid
...
    UpsertVaultWebhookRequest:
      type: object
      required:
        - url
      description: |
        There is no client-supplied secret field — the server always mints
        the signing secret.
      properties:
        url:
          type: string
          format: uri
          description: Must be an absolute https URL with a host.
          example: "https://hooks.example.com/rocketvault"
        rotate_secret:
          type: boolean
          description: Replace the existing signing secret with a newly minted one.
          default: false
        enabled:
          type: boolean
          description: Omitted leaves the current value unchanged (defaults to `true` on create).
    VaultWebhookConfigResponse:
      type: object
      description: |
        Has no property for the signing secret in any form, encrypted or
        plain — it is never returned by this shape.
      required:
        - url
        - enabled
        - created_at
        - updated_at
      properties:
        url:
          type: string
          format: uri
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VaultWebhookConfigCreatedResponse:
      type: object
      description: |
        VaultWebhookConfigResponse plus the plaintext signing secret, shown
        exactly once: only in the PUT response for the call that minted it
        (create, or rotate). No other endpoint or later call ever returns
        this field again.
      allOf:
        - $ref: "#/components/schemas/VaultWebhookConfigResponse"
        - type: object
          required:
            - signing_secret
          properties:
            signing_secret:
              type: string
              description: Plaintext signing secret. Shown once; not retrievable afterwards.

    # ─── Role assignments ────────────────────────────────────────────────────
...